+++ at_contacts_group_flutter/lib/screens/group_contact_view/group_contact_view.dart
@@ -34,7 +34,8 @@ class GroupContactView extends StatefulWidget {
   final bool asSelectionScreen;
 
   final bool isDesktop;
-  Function? onBackArrowTap, onDoneTap;
+  Function(List<GroupContactsModel?>?)? onBackArrowTap;
+  Function? onDoneTap;
 
   /// Callback to get the list of selected contacts back to the app
   final ValueChanged<List<GroupContactsModel?>>? selectedList;
@@ -111,8 +112,11 @@ class _GroupContactViewState extends State<GroupContactView> {
                         : Navigator.pop(context);
                   },
                   selectedList: (s) {
-                    widget.selectedList!(s);
+                    if (widget.selectedList != null) {
+                      widget.selectedList!(s);
+                    }
                   },
+                  isDesktop: widget.isDesktop,
                 )
               : Container(
                   height: 0,
@@ -121,7 +125,11 @@ class _GroupContactViewState extends State<GroupContactView> {
         isDesktop: widget.isDesktop,
         showTitle: true,
         titleText: 'Contacts',
-        onLeadingIconPressed: widget.onBackArrowTap,
+        onLeadingIconPressed: () {
+          if (widget.onBackArrowTap != null) {
+            widget.onBackArrowTap!(_groupService.selectedGroupContacts);
+          }
+        },
         showBackButton: true,
         showLeadingIcon: true,
         // showTrailingIcon: widget.asSelectionScreen == null ||