+++ at_contacts_group_flutter/lib/widgets/add_contacts_group_dialog.dart
@@ -111,8 +111,8 @@ class _AddContactDialogState extends State<AddContactDialog> {
                               setState(() {
                                 isLoading = true;
                               });
-                              var response = await _contactService
-                                  .addAtSign(context, atSign: atsignName);
+                              var response = await _contactService.addAtSign(
+                                  atSign: atsignName);
                               var _groupService = GroupService();
                               await _groupService.fetchGroupsAndContacts();
                               setState(() {